feat: implement IdeaDetail page with new components and error handling
