+++ widgets/idea-detail/index.ts
@@ -1,0 +1,6 @@
+export { IdeaDetailWidget } from './ui/IdeaDetailWidget';
+export { IdeaHeader } from './ui/IdeaHeader';
+export { IdeaContent } from './ui/IdeaContent';
+export { IdeaStats } from './ui/IdeaStats';
+export { LoadingState } from './ui/LoadingState';
+export { ErrorState } from './ui/ErrorState';
