+++ widgets/idea-detail/ui/LoadingState.tsx
@@ -1,0 +1,7 @@
+export function LoadingState() {
+  return (
+    <div className="min-h-screen bg-primary-palest flex items-center justify-center">
+      <div className="text-lg">載入中...</div>
+    </div>
+  );
+}
